+++ firmware/qemu/src/bin/timestamp.rs
@@ -1,9 +1,10 @@
 #![no_std]
 #![no_main]
 
+use cortex_m as _;
 use cortex_m_rt::entry;
-use cortex_m_semihosting::debug;
 use defmt::{write, Format, Formatter};
+use semihosting::process::ExitCode;
 
 use defmt_semihosting as _; // global logger
 
@@ -13,9 +14,7 @@ fn main() -> ! {
 
     defmt::println!("Hello {}{}", "World", '!');
 
-    loop {
-        debug::exit(debug::EXIT_SUCCESS)
-    }
+    ExitCode::SUCCESS.exit_process()
 }
 
 struct Timestamp {
@@ -44,7 +43,5 @@ defmt::timestamp!(
 #[cfg(target_os = "none")]
 #[panic_handler]
 fn panic(_: &core::panic::PanicInfo) -> ! {
-    loop {
-        debug::exit(debug::EXIT_FAILURE)
-    }
+    ExitCode::FAILURE.exit_process()
 }