+++ CHANGELOG.md
@@ -440,6 +440,7 @@ Initial release
 
 ### [defmt-macros-next]
 
+* [#1044]: Add support for `bound` attribute via a `#[defmt(bound(T: MyCustomBound))]`
 * [#956]: Link LICENSE-* in the crate folder
 
 ### [defmt-macros-v1.0.1] (2025-04-01)
@@ -1009,6 +1010,7 @@ Initial release
 [#1049]: https://github.com/knurling-rs/defmt/pull/1049
 [#1048]: https://github.com/knurling-rs/defmt/pull/1048
 [#1047]: https://github.com/knurling-rs/defmt/pull/1047
+[#1044]: https://github.com/knurling-rs/defmt/pull/1044
 [#1041]: https://github.com/knurling-rs/defmt/pull/1041
 [#1036]: https://github.com/knurling-rs/defmt/pull/1036
 [#1028]: https://github.com/knurling-rs/defmt/pull/1028